requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
alembic
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pytest
httpx
loguru
boto3
asyncpg
python-magic
